+++ src/Cooked/MockChain.hs
@@ -7,6 +7,7 @@ import Cooked.MockChain.Common as X
 import Cooked.MockChain.Effect.Misc as X
 import Cooked.MockChain.Effect.Read.Chain as X
 import Cooked.MockChain.Effect.Read.Conf as X
+import Cooked.MockChain.Effect.Time as X
 import Cooked.MockChain.Effect.Validation as X
 import Cooked.MockChain.Effect.Write as X
 import Cooked.MockChain.Run.Instances as X